+++ bpy/handlers/asciidoc.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# Copyright (C) 2013 by Yu-Jie Lin
+# Copyright (C) 2013, 2014 Yu-Jie Lin
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,17 @@
 # THE SOFTWARE.
 
 """
-No options are available at this moment.
+You can specify embed_images_, for example:
+
+.. code:: python
+
+  handlers = {
+    'AsciiDoc': {
+      'options': {
+        'embed_images': True,
+      },
+    },
+  }
 """
 
 from __future__ import print_function, unicode_literals